+++ publish.sh
@@ -9,28 +9,28 @@ sed -i '' "6s/.*/$var/" publish/manifest-firefox.json #sync version to other
 sed -i '' "6s/.*/$var/" publish/manifest-onprem.json
 sed -i '' "6s/.*/$var/" publish/manifest-firefox-onprem.json
 sed -i '' "6s/.*/$var/" publish/manifest-edge.json
-zip -r publish/snutils-chrome-base.zip . -x "*.DS_Store" -x "*.git*" -x ".jshintrc" -x ".docx" -x "*.sh" -x "*.md" -x "*publish*"
+zip -r publish/snutils-chrome-base.zip . -x "*.DS_Store" -x "*.git*" -x ".jshintrc" -x ".docx" -x "*.sh" -x "*.md" -x "*publish*" -x "*node_modules*"
 mv manifest.json publish/manifest-chrome.json
 
 sed -i '' "1s/.*/var onprem = true;/" background.js
 mv publish/manifest-onprem.json manifest.json
-zip -r publish/snutils-chrome-onprem.zip . -x "*.DS_Store" -x "*.git*" -x ".jshintrc" -x ".docx" -x "*.sh" -x "*.md" -x "*publish*"
+zip -r publish/snutils-chrome-onprem.zip . -x "*.DS_Store" -x "*.git*" -x ".jshintrc" -x ".docx" -x "*.sh" -x "*.md" -x "*publish*" -x "*node_modules*"
 mv manifest.json publish/manifest-onprem.json
 
 mv publish/manifest-firefox.json manifest.json
 mv js/monaco/vs/language/typescript/tsWorker.js publish/tsWorker.js 
 mv publish/tsWorkerFF.js js/monaco/vs/language/typescript/tsWorker.js
-zip -r publish/snutils-firefox-base.xpi . -x "*.DS_Store" -x "*.git*" -x ".jshintrc" -x ".docx" -x "*.sh" -x "*.md" -x "*publish*"
+zip -r publish/snutils-firefox-base.xpi . -x "*.DS_Store" -x "*.git*" -x ".jshintrc" -x ".docx" -x "*.sh" -x "*.md" -x "*publish*" -x "*node_modules*"
 mv manifest.json publish/manifest-firefox.json
 
 mv publish/manifest-firefox-onprem.json manifest.json
-zip -r publish/snutils-firefox-onprem.xpi . -x "*.DS_Store" -x "*.git*" -x ".jshintrc" -x ".docx" -x "*.sh" -x "*.md" -x "*publish*"
+zip -r publish/snutils-firefox-onprem.xpi . -x "*.DS_Store" -x "*.git*" -x ".jshintrc" -x ".docx" -x "*.sh" -x "*.md" -x "*publish*" -x "*node_modules*"
 mv js/monaco/vs/language/typescript/tsWorker.js publish/tsWorkerFF.js
 mv publish/tsWorker.js js/monaco/vs/language/typescript/tsWorker.js
 mv manifest.json publish/manifest-firefox-onprem.json
 
 mv publish/manifest-edge.json manifest.json
-zip -r publish/snutils-edge-base.zip . -x "*.DS_Store" -x "*.git*" -x ".jshintrc" -x ".docx" -x "*.sh" -x "*.md" -x "*publish*"
+zip -r publish/snutils-edge-base.zip . -x "*.DS_Store" -x "*.git*" -x ".jshintrc" -x ".docx" -x "*.sh" -x "*.md" -x "*publish*" -x "*node_modules*"
 mv manifest.json publish/manifest-edge.json
 
 mv publish/manifest-chrome.json manifest.json